+++ lib/simplegamelib.dart
@@ -17,3 +17,4 @@ export 'src/unimplemented.dart';
 export 'src/game.dart';
 export 'src/movements.dart';
 export 'src/audiobank.dart';
+export 'src/collectible.dart';