+++ lib/src/audiobank.dart
@@ -5,13 +5,11 @@ import 'dart:html';
 /// Audio Utility Class
 /// Note : Dartium will NOT play MP3 files.
 class AudioBank {
-
-  Map<String,AudioElement> sounds = {};
+  Map<String, AudioElement> sounds = {};
 
   /// Play the sound with the given name.
   void play(String name) {
-    if (sounds.containsKey(name))
-      sounds[name].play();
+    if (sounds.containsKey(name)) sounds[name].play();
   }
 
   /// Load a sound into the audio bank.