+++ lib/src/player.dart
@@ -1,4 +1,5 @@
 library simplegamelib.player;
+
 import 'sprite.dart';
 
 /// Active agent taking part in the game.
@@ -9,11 +10,11 @@ class Player {
   int score;
   String name;
 
-  Player(){
+  Player() {
     reset();
   }
 
-  void  reset(){
+  void reset() {
     health = 100;
     score = 0;
     name = "Player1";