+++ lib/src/unimplemented.dart
@@ -1,12 +1,10 @@
 library simplegamelib.unimplemented;
+
 /// Limits a sprite to a particular area.
-class Limiter {
-}
+class Limiter {}
 
 /// A compendium of images and sounds.
-class ResourcePack {
-}
+class ResourcePack {}
 
 /// Play area object.
-class Arena {
-}
+class Arena {}